scripts/minio-setup/requirements.txt
boto3>=1.34.0
minio>=7.2.0
python-dotenv>=1.0.0
click>=8.1.7
